pyproject.toml
[project]
name = "ticketbester"
version = "0.1.0"
description = "The best ticketing app you'll see in your life."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "pyqt6-tools>=6.4.2.3.3",
    "pyside6>=6.10.1",
    "segno>=1.6.6",
]
